# uint256: optimize Mul, squared and Exp